requirements.txt
aiogram
sqlalchemy
pydantic_settings
aiofiles
pillow
alembic
aiogram_dialog
asyncpg
